pyproject.toml
[tool.poetry]
name = "reposage"
version = "0.1.0"
description = "Code intelligence tool for ObjC/Swift/Java — dual-layer docs + MCP server"
authors = []
readme = "README.md"
packages = [{include = "reposage"}]

[tool.poetry.dependencies]
python = ">=3.9"
mcp = ">=1.0.0"
anthropic = ">=0.40.0"
tree-sitter = ">=0.21.0,<0.22.0"
tree-sitter-languages = ">=1.10.0"
chromadb = ">=0.5.0"
watchdog = ">=4.0.0"
fastapi = ">=0.110.0"
uvicorn = ">=0.29.0"
pydantic = ">=2.0.0"
pyyaml = ">=6.0"
click = ">=8.0.0"
rich = ">=13.0.0"

[tool.poetry.scripts]
reposage = "reposage.cli:cli"

[tool.poetry.dev-dependencies]
pytest = ">=7.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
